Fixed mass slider issue on all search pages

+++ Front-End/src/js/glycan_search.js
@@ -144,9 +144,9 @@ Sliderbox.prototype.handler = function (target) {
     noUiSlider.create(slider, this.options);
     slider.noUiSlider.on('update', function (values, handle) {
         if (handle) {
-            inpMax.value = addCommas(parseInt(values[handle]));
+            inpMax.value = addCommas(parseFloat(values[handle]));
         } else {
-            inpMin.value = addCommas(parseInt(values[handle]));
+            inpMin.value = addCommas(parseFloat(values[handle]));
         }
     });
 
@@ -368,8 +368,8 @@ function searchjson(input_query_type, input_glycan_id, mass_min, mass_max, sugar
         "operation": "AND",
         query_type: input_query_type,
         mass: {
-            "min": parseInt(mass_min),
-            "max": parseInt(mass_max)
+            "min": parseFloat(mass_min),
+            "max": parseFloat(mass_max)
         },
         // mass:masses,
         number_monosaccharides: {
